feat: 밴드_삭제_시_자신이_생성한_밴드가_아니면_UNAUTHORIZED_BAND_예외를_반환한다

+++ wabi/src/main/kotlin/com/wap/wabi/band/service/BandService.kt
@@ -48,6 +48,10 @@ class BandService(
             throw RestApiException(ErrorCode.ALREADY_ADD_STUDENT)
         }
 
+        if (band.adminId != adminId) {
+            throw RestApiException(ErrorCode.UNAUTHORIZED_BAND)
+        }
+
         bandRepository.delete(band)
     }
